+++ core/dataStores/index.js
@@ -1,9 +1,0 @@
-'use strict';
-
-const StorageMemory = require('./storageMemory');
-const StorageDataGroup = require('./storageDataGroup');
-
-module.exports = {
-    StorageMemory,
-    StorageDataGroup
-};