+++ core/test/template.js
@@ -45,6 +45,9 @@ describe('Template class tests', function () {
         const reference = new Template();
         reference.description = 'Just a basic template';
         reference.templateText = mstdata;
+        reference.sourceType = 'MST';
+        reference.sourceText = mstdata;
+        reference.sourceHash = '6ac8bbb53fdfe637931e0dfc9e4259ef685ab5fc8e1e13b796dbf6d3145fe213';
 
         return Template.loadMst(mstdata)
             .then((tmpl) => {
@@ -69,9 +72,27 @@ describe('Template class tests', function () {
               </html>
         `;
 
+        const reference = new Template();
+        reference.description = '';
+        reference.defaultView = {
+            message: 'Hello!'
+        };
+        reference.definitions = {
+            body: {
+                template: '<body> <h1>{{message}}</h1> </body>'
+            }
+        };
+        reference.templateText = '<html>\n  {{> body}}\n</html>\n';
+        reference.sourceType = 'YAML';
+        reference.sourceText = ymldata;
+        reference.sourceHash = '99223c057171b3aceb955c953a8efdf237e65e36b287138ecd953585217fe783';
+
         return Template.loadYaml(ymldata)
             .then((tmpl) => {
                 assert.ok(tmpl);
+
+                reference._viewSchema = tmpl._viewSchema;
+                assert.deepStrictEqual(tmpl, reference);
             });
     });
     it('from_json', function () {
@@ -211,7 +232,7 @@ describe('Template class tests', function () {
                 const schema = tmpl.getViewSchema();
                 console.log(JSON.stringify(schema, null, 2));
                 assert.strictEqual(tmpl.render(view), reference);
-                assert.strictEqual(schema.properties.foo.invertDependency, true);
+                assert.deepStrictEqual(schema.properties.foo.invertDependency, ['skip_foo']);
             });
     });
     it('render_type_defaults', function () {
@@ -290,6 +311,28 @@ describe('Template class tests', function () {
             .then((tmpl) => {
                 console.log(JSON.stringify(tmpl.getViewSchema(), null, 2));
                 assert.strictEqual(tmpl.render(view).trim(), reference);
+            });
+    });
+    it('schema_nested_sections', function () {
+        const ymldata = `
+            definitions:
+                part:
+                    template: |
+                        {{^use_existing_a}}
+                            {{^make_new_b}}
+                                {{value}}
+                            {{/make_new_b}}
+                        {{/use_existing_a}}
+            template: |
+                {{> part}}
+        `;
+        return Template.loadYaml(ymldata)
+            .then((tmpl) => {
+                assert.ok(tmpl);
+
+                const schema = tmpl.getViewSchema();
+                console.log(JSON.stringify(schema, null, 2));
+                assert.deepStrictEqual(schema.dependencies.value, ['use_existing_a', 'make_new_b']);
             });
     });
     it('schema_clean_deps', function () {