+++ iappslx/presentation/app.js
@@ -1,10 +1,11 @@
+/* eslint-env browser */
+
 'use strict';
+
 
 const yaml = require('js-yaml');
 
 const { Template, guiUtils } = require('@f5devcentral/fast');
-const SwaggerUI = require('swagger-ui');
-const specData = require('./openapi.json');
 
 const endPointUrl = '/mgmt/shared/fast';
 
@@ -24,6 +25,42 @@ const dispOutput = (output) => {
     if (outputElem) {
         outputElem.innerText = output;
     }
+};
+
+const multipartUpload = (file) => {
+    const CHUNK_SIZE = 1000000;
+    const uploadPart = (start, end) => {
+        const opts = {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/octet-stream',
+                'Content-Range': `${start}-${end}/${file.size}`,
+                'Content-Length': (end - start) + 1,
+                'Connection': 'keep-alive'
+            },
+            body: file.slice(start, end + 1, 'application/octet-stream')
+        };
+        return fetch(`/mgmt/shared/file-transfer/uploads/${file.name}`, opts)
+            .then((response) => {
+                if (!response.ok) {
+                    return Promise.reject(new Error(`Failed to upload file: ${JSON.stringify(response)}`));
+                }
+                return Promise.resolve();
+            })
+            .then(() => {
+                if (end === file.size - 1) {
+                    return Promise.resolve();
+                }
+                const nextStart = start + CHUNK_SIZE;
+                const nextEnd = (end + CHUNK_SIZE > file.size - 1) ? file.size - 1 : end + CHUNK_SIZE;
+                return uploadPart(nextStart, nextEnd);
+            });
+    };
+
+    if (CHUNK_SIZE < file.size) {
+        return uploadPart(0, CHUNK_SIZE-1);
+    }
+    return uploadPart(0, file.size-1);
 };
 
 
@@ -76,10 +113,13 @@ const newEditor = (tmplid, view) => {
             // Hook up buttons
             document.getElementById('view-tmpl-btn').onclick = () => {
                 dispOutput(tmpl.templateText);
-            }
+            };
             document.getElementById('view-schema-btn').onclick = () => {
                 dispOutput(JSON.stringify(schema, null, 2));
-            }
+            };
+            document.getElementById('view-view-btn').onclick = () => {
+                dispOutput(JSON.stringify(tmpl.getCombinedView(editor.getValue()), null, 2));
+            };
             document.getElementById('view-render-btn').onclick = () => {
                 dispOutput(JSON.stringify(yaml.safeLoad(tmpl.render(editor.getValue())), null, 2));
             };
@@ -112,7 +152,7 @@ function route(path, pageName, pageFunc) {
     routes[path] = { pageName, pageFunc };
 }
 
-const navTitles = {}
+const navTitles = {};
 function addRouteToHeader(topRoute, title) {
     navTitles[topRoute] = title;
     return title;
@@ -134,7 +174,7 @@ function router() {
             d = document.createElement('a');
             d.classList.add('btn-nav');
             d.classList.add('btn');
-            d.href = `#${k}`
+            d.href = `#${k}`;
         } else {
             d = document.createElement('div');
             d.classList.add('selected-nav');
@@ -182,7 +222,7 @@ addRouteToHeader('api', 'API');
 
 // Define routes
 route('', 'apps', () => {
-    outputElem =  document.getElementById('output');
+    outputElem = document.getElementById('output');
     const listElem = document.getElementById('applist');
     let count = 0;
     let lastTenant = '';
@@ -215,7 +255,7 @@ route('', 'apps', () => {
                 appTenant.classList.add('appListTitle');
                 row.appendChild(appTenant);
 
-                const appName= document.createElement('div');
+                const appName = document.createElement('div');
                 appName.innerText = appPair[1];
                 appName.classList.add('appListTitle');
                 row.appendChild(appName);
@@ -241,14 +281,13 @@ route('', 'apps', () => {
                 deleteBtn.addEventListener('click', () => {
                     dispOutput(`Deleting ${appPairStr}`);
                     fetch(`${endPointUrl}/applications/${appPairStr}`, {
-                        method: 'DELETE',
+                        method: 'DELETE'
                     });
                 });
                 deleteBtn.classList.add('appListEntry');
                 row.appendChild(deleteBtn);
-
-
             });
+
             dispOutput('');
         })
         .catch(e => dispOutput(`Error fetching applications: ${e.message}`));
@@ -268,7 +307,7 @@ route('create', 'create', () => {
             elem.appendChild(btn);
         });
     };
-    outputElem =  document.getElementById('output');
+    outputElem = document.getElementById('output');
     dispOutput('Fetching templates');
     getJSON('templates')
         .then((data) => {
@@ -278,7 +317,7 @@ route('create', 'create', () => {
         .catch(e => dispOutput(e.message));
 });
 route('modify', 'create', (appID) => {
-    outputElem =  document.getElementById('output');
+    outputElem = document.getElementById('output');
     dispOutput(`Fetching app data for ${appID}`);
     getJSON(`applications/${appID}`)
         .then((appData) => {
@@ -289,98 +328,142 @@ route('modify', 'create', (appID) => {
 });
 
 route('tasks', 'tasks', () => {
-  console.log('Fetching AS3 Tasks Endpoint');
-    fetch('/mgmt/shared/appsvcs/task')
-        .then((data) => {
-            return data.json();
-        })
-        .then((data)=> {
-            const taskList = document.getElementById('task-list');
-            taskList.innerHTML = `<div class="appListRow">
-              <div class="appListTitle">Task ID</div>
-              <div class="appListTitle">Tenant</div>
-              <div class="appListTitle">Result</div>
-            </div>`
-            taskList.appendChild(document.createElement('hr'));
-
-            let count = 0;
-            data.items.forEach((item) => {
-                const rowDiv = document.createElement('div');
-                rowDiv.classList.add('appListRow');
-                if (++count%2) rowDiv.classList.add('zebraRow');
-
-                const idDiv = document.createElement('div');
-                idDiv.classList.add('appListTitle');
-                idDiv.innerText = item.id;
-                rowDiv.appendChild(idDiv);
-
-                const tenantDiv = document.createElement('div');
-                tenantDiv.classList.add('appListTitle');
-
-                const statusDiv = document.createElement('div');
-                statusDiv.classList.add('appListTitle');
-
-                const changes = item.results.filter((r) => r.message !== 'no change');
-                if (changes.length === 0) {
-                    statusDiv.innerText = 'no change';
-                } else {
-                    changes.forEach((change) => {
-                        const tDiv = document.createElement('div');
-                        tDiv.innerText = `${change.tenant}`
-                        const mDiv = document.createElement('div');
-                        mDiv.innerText = `${change.message}`;
-                        if (change.errors) {
-                            mDiv.innerText += `\n${change.errors.join('\n')}`;
-                        }
-                        if (change.response) {
-                            mDiv.innerText += `\n${change.response}`;
-                        }
-                        tenantDiv.appendChild(tDiv);
-                        statusDiv.appendChild(mDiv);
-                    });
-                }
-
-                rowDiv.appendChild(tenantDiv);
-                rowDiv.appendChild(statusDiv);
-
-                taskList.appendChild(rowDiv);
+    const renderTaskList = () => {
+        fetch('/mgmt/shared/appsvcs/task')
+            .then(data => data.json())
+            .then((data) => {
+                const taskList = document.getElementById('task-list');
+                taskList.innerHTML = `<div class="appListRow">
+                  <div class="appListTitle">Task ID</div>
+                  <div class="appListTitle">Tenant</div>
+                  <div class="appListTitle">Result</div>
+                </div>`;
                 taskList.appendChild(document.createElement('hr'));
-            });
 
-        });
+                let count = 0;
+                data.items.forEach((item) => {
+                    const rowDiv = document.createElement('div');
+                    rowDiv.classList.add('appListRow');
+                    if (++count%2) rowDiv.classList.add('zebraRow');
+
+                    const idDiv = document.createElement('div');
+                    idDiv.classList.add('appListTitle');
+                    idDiv.innerText = item.id;
+                    rowDiv.appendChild(idDiv);
+
+                    const tenantDiv = document.createElement('div');
+                    tenantDiv.classList.add('appListTitle');
+
+                    const statusDiv = document.createElement('div');
+                    statusDiv.classList.add('appListTitle');
+
+                    const changes = item.results.filter(r => r.message !== 'no change');
+                    if (changes.length === 0) {
+                        statusDiv.innerText = 'no change';
+                    } else {
+                        changes.forEach((change) => {
+                            const tDiv = document.createElement('div');
+                            tDiv.innerText = `${change.tenant}`;
+                            const mDiv = document.createElement('div');
+                            mDiv.innerText = `${change.message}`;
+                            if (change.errors) {
+                                mDiv.innerText += `\n${change.errors.join('\n')}`;
+                            }
+                            if (change.response) {
+                                mDiv.innerText += `\n${change.response}`;
+                            }
+                            tenantDiv.appendChild(tDiv);
+                            statusDiv.appendChild(mDiv);
+                        });
+                    }
+
+                    rowDiv.appendChild(tenantDiv);
+                    rowDiv.appendChild(statusDiv);
+
+                    taskList.appendChild(rowDiv);
+                    taskList.appendChild(document.createElement('hr'));
+                });
+
+                const inProgressJob = (
+                    data.items[0] &&
+                    data.items[0].results &&
+                    data.items[0].results[0] &&
+                    data.items[0].results[0].message === 'in progress'
+                );
+                if (inProgressJob) {
+                    setTimeout(renderTaskList, 5000);
+                }
+            });
+    };
+
+    renderTaskList();
 });
 
 route('api', 'api', () => {
-    SwaggerUI({
-        dom_id: '#api-div',
-        spec: specData
-    });
+
 });
 
 route('templates', 'templates', () => {
-    console.log('Fetching Template Table Data.');
+    outputElem = document.getElementById('output');
+    dispOutput('...Loading Template List...');
     const templateDiv = document.getElementById('template-list');
-    Promise.all([getJSON('applications'),
-        getJSON('templates')])
-        .then((data) => {
-            templateDiv.innerHTML = `<div class="appListRow">
-              <div class="appListTitle">Templates</div>
-              <div class="appListTitle">Applications</div>
-            </div>`;
-            const applications = data[0];
-            const templates = data[1];
+
+    document.getElementById('add-ts-btn').onclick = () => {
+        document.getElementById('ts-file-input').click();
+    };
+    document.getElementById('ts-file-input').onchange = (input) => {
+        const file = document.getElementById('ts-file-input').files[0]
+        const tsName = file.name.slice(0, -4);
+        dispOutput(`Uploading file: ${file.name}`);
+        multipartUpload(file)
+            .then(() => dispOutput(`Installing template set ${tsName}`))
+            .then(() => fetch('/mgmt/shared/fast/templatesets', {
+                method: 'POST',
+                headers: {
+                    'Content-Type': 'application/json'
+                },
+                body: JSON.stringify({
+                    name: file.name.slice(0, -4)
+                })
+            }))
+            .then((response) => {
+                if (!response.ok) {
+                    return Promise.reject(new Error(`Failed to install ${tsName}: ${JSON.stringify(response)}`));
+                }
+                return Promise.resolve();
+            })
+            .then(() => dispOutput(`${tsName} installed successfully`))
+            .then(() => location.reload())
+            .catch(e => dispOutput(`Failed to install template set: ${e.stack}`));
+    };
+
+    Promise.all([
+        getJSON('applications'),
+        getJSON('templates'),
+        getJSON('templatesets')
+    ])
+        .then(([applications, templates, templatesets]) => {
+            const setMap = templatesets.reduce((acc, curr) => {
+                acc[curr.name] = curr;
+                return acc;
+            }, {});
 
             // build dictionary of app lists, keyed by template
             const appDict = applications.reduce((a, c) => {
                 if (c.template) {
-                    if(!a[c.template])
+                    if (!a[c.template]) {
                         a[c.template] = [];
+                    }
                     a[c.template].push(c);
                 }
                 return a;
             }, {});
+
             let count = 0;
-            templates.forEach((tname) => {
+            const createRow = (rowName, rowList, actionsList) => {
+                rowList = rowList || [];
+                actionsList = actionsList || [];
+
                 const row = document.createElement('div');
                 row.classList.add('appListRow');
                 if (++count%2) {
@@ -388,24 +471,71 @@ route('templates', 'templates', () => {
                 }
 
                 const name = document.createElement('div');
-                name.innerText = tname;
+                name.innerHTML = rowName;
                 name.classList.add('appListTitle');
                 row.appendChild(name);
 
                 const applist = document.createElement('div');
                 applist.classList.add('appListTitle');
-                if (appDict[tname]) {
-                    appDict[tname].forEach((app) => {
-                        const appDiv = document.createElement('div');
-                        appDiv.innerText = app.tenant + ' ' + app.name;
-                        applist.appendChild(appDiv);
-                    });
-                }
+                rowList.forEach((item) => {
+                    const appDiv = document.createElement('div');
+                    appDiv.innerText = item;
+                    applist.appendChild(appDiv);
+                });
                 row.appendChild(applist);
+
+                const actions = document.createElement('div');
+                actions.classList.add('appListTitle');
+                Object.entries(actionsList).forEach(([actName, actFn]) => {
+                    const actBtn = document.createElement('a');
+                    actBtn.classList.add('btn');
+                    actBtn.innerText = actName;
+                    actBtn.onclick = actFn;
+                    actions.appendChild(actBtn);
+                });
+                row.appendChild(actions);
 
                 templateDiv.appendChild(document.createElement('hr'));
                 templateDiv.appendChild(row);
+            };
+            Object.values(setMap).forEach((setData) => {
+                const setName = setData.name;
+                const setActions = {
+                    Remove: () => {
+                        dispOutput(`Deleting ${setName}`);
+                        return fetch(`${endPointUrl}/templatesets/${setName}`, {
+                            method: 'DELETE'
+                        })
+                            .then(() => location.reload());
+                    }
+                };
+
+                if (setData.updateAvailable) {
+                    setActions.Update = () => {
+                        dispOutput(`Updating ${setName}`);
+                        return fetch(`${endPointUrl}/templatesets/${setName}`, {
+                            method: 'DELETE'
+                        })
+                            .then(() => fetch(`${endPointUrl}/templatesets`, {
+                                method: 'POST',
+                                headers: {
+                                    'Content-Type': 'application/json'
+                                },
+                                body: JSON.stringify({
+                                    name: setName
+                                })
+                            }))
+                            .then(() => location.reload());
+                    };
+                }
+
+                createRow(setName, (setData.supported) ? ['supported'] : [], setActions);
+                setMap[setName].templates.forEach((tmpl) => {
+                    const templateName = tmpl.name;
+                    const appList = appDict[`${setName}/${templateName}`] || [];
+                    createRow(`&nbsp;&nbsp;&nbsp;&nbsp;/${templateName}`, appList.map(app => `${app.tenant} ${app.name}`));
+                });
             });
-            templateDiv.appendChild(document.createElement('hr'));
-        });
+        })
+        .then(() => dispOutput(''));
 });